Adding support for splicing (#6)
* WIP branch for working on spackmonitor
* version bump
* small tweak to logic program
* adding support for splicing
* add runtest function directly to file
* add notes for how to run tests

+++ examples/cpp/MathClient.cpp
@@ -7,7 +7,7 @@
 int main()
 {
     double a = 7.4;
-    int b = 99;
+    double b = 99;
 
     std::cout << "a + b = " <<
         MathLibrary::Arithmetic::Add(a, b) << std::endl;
